requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aioredis==2.0.1
aiormq==3.3.1
alembic==1.4.3
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
bidict==0.22.0
boto3==1.21.29
botocore==1.24.29
CacheControl==0.12.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.1
clipboard==0.0.4
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
croniter==1.3.4
cryptography==3.4.8
cycler==0.11.0
dask==2021.11.2
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
fasttext==0.9.2
fbmessenger==6.0.0
fire==0.4.0
flatbuffers==2.0
fsspec==2022.2.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
httptools==0.4.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
jmespath==1.0.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
locket==0.2.1
lxml==4.6.5
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.3
multidict==5.2.0
networkx==2.5.1
nltk==3.7
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.9
pamqp==2.3.0
partd==1.2.0
passlib==1.7.4
pika==1.2.0
Pillow==9.0.1
pip==22.0.4
pipreqs==0.4.11
prompt-toolkit==2.0.10
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
pydot==1.4.2
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline==2.1
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-engineio==4.3.1
python-socketio==5.5.2
python3-saml==1.14.0
pytz==2021.3
pyvi==0.1.1
PyYAML==6.0
questionary==1.5.2
randomname==0.1.5
rasa==3.1.0
rasa-sdk==3.1.1
rasa-x==1.1.0
redis==3.5.3
regex==2021.8.28
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sanic==21.12.1
Sanic-Cors==2.0.1
sanic-jwt==1.7.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
selenium==4.1.3
sentry-sdk==1.3.1
setuptools==61.2.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.1
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
toolz==0.11.2
tqdm==4.63.1
trio==0.20.0
trio-websocket==0.9.2
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==1.35
underthesea==1.3.4
underthesea_core==0.0.4a10
Unidecode==1.3.4
urllib3==1.26.9
wcwidth==0.2.5
webexteamssdk==1.6
websockets==10.2
Werkzeug==2.1.0
wheel==0.37.1
wrapt==1.14.0
wsproto==1.1.0
xmlsec==1.3.12
yarg==0.1.9
yarl==1.7.2
zipp==3.7.0
